Feat: V2 endpoints

+++ src/main/java/com/example/api/entity/Transaction.java
@@ -6,6 +6,9 @@ import lombok.*;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 
+/**
+ * Class responsible for the structure of the transaction entity, mapped to the table "transactions".
+ */
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -20,10 +23,10 @@ public class Transaction implements Serializable {
     //@ManyToOne
     //@JoinColumn(name = "to_account_number")
 
-    private Long toAccountNumber;
+    private Integer toAccountNumber;
     //@ManyToOne
     //@JoinColumn(name = "from_account_number")
-    private Long fromAccountNumber;
+    private Integer fromAccountNumber;
     @Column(name = "transaction_value")
     private Double value;
     @Column(name = "transaction_date")
